Protect flymake-ui against commands like fill-paragraph
If flymake-start-syntax-check-on-newline is t, check should start as
soon as a newline is seen by after-change-functions.  But don't rush
it because since the buffer state might not be final and we end up
with invalid diagnostic regions after some commands, like
fill-paragraph, which might insert and deletes newlines.

* lisp/progmodes/flymake-ui.el
(flymake-after-change-function): Pass `deferred' to
flymake--start-syntax-check.
(flymake--start-syntax-check): Take optional `deferred' arg.