Allow disable the record of keystrokes (lossage)
Use 1 as the minimum value for lossage-limit; such a value
is equivalent to not recording the keystrokes: having just 1 entry,
will be overwritten with the view-lossage call itself.

* test/src/keyboard-tests.el (keyboard-lossage-limit): Update test.
* src/keyboard.c (MIN_NUM_RECENT_KEYS): Delete it.
(lossage_limit): Add security note in the doctring.
* lisp/cus-start.el (lossage-limit): Let users choose to disable
the record of the keystrokes.
* doc/emacs/help.texi (Misc Help): Update manual.
* etc/NEWS (Changes in Emacs 28.1):
Mention that it's possible to disable the record of keystrokes.